+++ src/backends/webgpu/src/backend_webgpu.ts
@@ -191,6 +191,7 @@ export class WebGPUBackend extends KernelBackend {
       // vec3 and vec4 have the same alignment, however padding is only
       // necessary in case of adjacent vec3's. Complete std140 layout rules are
       // documented here:
+      // tslint:disable-next-line:max-line-length
       // https://www.khronos.org/registry/OpenGL/specs/gl/glspec45.core.pdf#page=159
       if (d.length === 3 && i > 0 && bufferShapes[i - 1].length === 3) {
         dimUniforms.push(0);
