+++ src/backends/webgpu/src/shader_util.ts
@@ -15,8 +15,6 @@
  * =============================================================================
  */
 
-import {getCoordsDataType} from './shader_preprocessor';
-
 // Generates GLSL that computes strides.
 export function symbolicallyComputeStrides(
     indicesArr: number[], variableName: string): string[] {
@@ -33,54 +31,4 @@ export function symbolicallyComputeStrides(
   }
 
   return strides;
-}
-
-/**
- * Generates getOutputCoords() function that computes output coordinates from
- * dispatch geometry to reduce arithmetic.
- */
-export function generateGetOutputCoords(
-    dispatchLayout: {x: number[], y?: number[], z?: number[]},
-    rank: number): string {
-  const {x, y = [], z = []} = dispatchLayout;
-  const dtype = getCoordsDataType(rank);
-  let gatherDimensionsStr = '';
-  const dims = [x, y, z];
-
-  for (let i = 0; i < dims.length; i++) {
-    const arr = dims[i];
-
-    if (arr.length === 0) {
-      continue;
-    }
-
-    if (arr.length === 1) {
-      gatherDimensionsStr += `uint d${arr[0]} = gl_GlobalInvocationID[${i}];`;
-    } else {
-      const strides = symbolicallyComputeStrides(arr, 'outShape');
-      gatherDimensionsStr += `uint index${i} = 
-            gl_GlobalInvocationID[${i}];`;
-      for (let j = 0; j < strides.length; j++) {
-        gatherDimensionsStr += `uint d${arr[j]} = index${i} / ${strides[j]};`;
-
-        if (j === strides.length - 1) {
-          gatherDimensionsStr += `uint d${arr[j + 1]} = ` +
-              `index${i} - d${arr[j]} * ${strides[j]};`;
-        } else {
-          gatherDimensionsStr += `index${i} -= d${arr[j]} * ${strides[j]};`;
-        }
-      }
-    }
-  }
-
-  const dimensions = [];
-  for (let i = 0; i < rank; i++) {
-    dimensions.push(`d${i}`);
-  }
-
-  return `${dtype} getOutputCoords() {
-    ${gatherDimensionsStr}
-
-    return ${dtype}(${dimensions.join(',')});
-  }`;
 }
